fix: improve code quality
