+++ backend/collaboration-service/src/services/yjs.service.js
@@ -38,7 +38,6 @@ server.on('upgrade', (request, socket, head) => {
 wss.on('connection', (ws, req, user) => {
     const docName = req.url.slice(1) // Use the URL as the document name
     setupWSConnection(ws, req, { docName })
-    console.log(`User ${user.id} connected to document: ${docName}`)
     ws.on('close', () => {
         console.log(`User ${user.id} disconnected`)
     })
