+++ frontend/types/editor.ts
@@ -1,0 +1,4 @@
+export interface CodeMirrorEditorProps {
+    roomId: string
+    language: string
+}
